Refactor: move pistesyotto table to its own subfolder
