+++ src/app/haku/[oid]/hakukohde/[hakukohde]/pistesyotto/table/koe-cell.tsx
@@ -1,9 +1,8 @@
 'use client';
-import { OphSelect } from '@/app/components/form/oph-select';
-import ListTable, {
-  makeColumnWithCustomRender,
-  makeExternalLinkColumn,
-} from '@/app/components/table/list-table';
+import {
+  InputValidator,
+  numberValidator,
+} from '@/app/components/form/input-validators';
 import { useTranslations } from '@/app/hooks/useTranslations';
 import {
   HakemuksenPistetiedot,
@@ -14,27 +13,15 @@ import {
   Valintakoe,
   ValintakoeInputTyyppi,
 } from '@/app/lib/types/valintaperusteet-types';
-import { Box, SelectChangeEvent, Typography } from '@mui/material';
 import { ChangeEvent, useState } from 'react';
 import { AnyEventObject } from 'xstate';
-import { PisteSyottoEvents } from './pistesyotto-state';
-import { colors } from '@opetushallitus/oph-design-system';
-import {
-  isNotPartOfThisHakukohde,
-  NOT_READABLE_REASON_MAP,
-} from './pistesyotto-utils';
-import {
-  numberValidator,
-  InputValidator,
-} from '@/app/components/form/input-validators';
+import { PisteSyottoEvents } from '../pistesyotto-state';
+import { Box, SelectChangeEvent } from '@mui/material';
+import { OphSelect } from '@/app/components/form/oph-select';
 import { OphFormControl } from '@/app/components/form/oph-form-control';
 import { OphInput } from '@/app/components/form/oph-input';
 
-const LINK_TO_PERSON = 'henkilo-ui/oppija/';
-
-const buildLinkToPerson = (personOid: string) => LINK_TO_PERSON + personOid;
-
-const KoeCell = ({
+export const KoeCell = ({
   pisteTiedot,
   send,
   koe,
@@ -185,102 +172,5 @@ const KoeCell = ({
         sx={{ minWidth: '150px' }}
       />
     </Box>
-  );
-};
-
-const ReadOnlyCell = ({
-  pisteTiedot,
-  koe,
-}: {
-  pisteTiedot: HakemuksenPistetiedot;
-  koe: Valintakoe;
-}) => {
-  const { t } = useTranslations();
-
-  const pisteet = pisteTiedot.valintakokeenPisteet.find(
-    (vp) => vp.tunniste === koe.tunniste,
-  );
-  const notReadableReason = pisteet?.osallistuminen
-    ? NOT_READABLE_REASON_MAP[pisteet?.osallistuminen]
-    : '';
-
-  return (
-    <Box
-      sx={{
-        display: 'flex',
-        flexDirection: 'row',
-        columnGap: '0.6rem',
-        minWidth: '220px',
-      }}
-    >
-      <Typography>{pisteet?.arvo}</Typography>
-      <Typography>{t(notReadableReason)}</Typography>
-    </Box>
-  );
-};
-
-const stickyColumnStyle: React.CSSProperties = {
-  minWidth: '200px',
-  position: 'sticky',
-  left: 0,
-  borderRight: `2px solid ${colors.grey100}`,
-  zIndex: 1,
-  backgroundColor: colors.white,
-};
-
-export const PisteSyottoTable = ({
-  pistetiedot,
-  setSort,
-  sort,
-  kokeet,
-  send,
-}: {
-  pistetiedot: HakemuksenPistetiedot[];
-  sort: string;
-  setSort: (sort: string) => void;
-  kokeet: Valintakoe[];
-  send: (event: AnyEventObject) => void;
-}) => {
-  const { t } = useTranslations();
-
-  const hakijaColumn = Object.assign(
-    makeExternalLinkColumn<HakemuksenPistetiedot>({
-      linkBuilder: buildLinkToPerson,
-      title: t('hakeneet.taulukko.hakija'),
-      key: 'hakijanNimi',
-      nameProp: 'hakijanNimi',
-      linkProp: 'hakijaOid',
-    }),
-    { style: stickyColumnStyle },
-  );
-
-  const koeColumns = kokeet.map((koe) => {
-    return makeColumnWithCustomRender<HakemuksenPistetiedot>({
-      title: koe.kuvaus,
-      key: koe.tunniste,
-      renderFn: (props) =>
-        isNotPartOfThisHakukohde(
-          props.valintakokeenPisteet.find((p) => p.tunniste === koe.tunniste)!,
-        ) ? (
-          <ReadOnlyCell pisteTiedot={props} koe={koe} />
-        ) : (
-          <KoeCell pisteTiedot={props} koe={koe} send={send} />
-        ),
-      sortable: false,
-    });
-  });
-
-  const columns = [hakijaColumn, ...koeColumns];
-
-  return (
-    <ListTable
-      rowKeyProp="hakemusOid"
-      columns={columns}
-      rows={pistetiedot}
-      sort={sort}
-      setSort={setSort}
-      translateHeader={false}
-      sx={{ overflowX: 'auto', width: 'unset' }}
-    />
   );
 };
